Improve the ozman sieve
I found a way to simplify the Ozman sieve computationally, also in the more general case where N is no longer assumed to be prime. One of course still needs the complete splitting condition in the Hilbert class field. But the existence of a residue field 1 degree one prime above p in B can be made easier. Namely we just need p to split in all real quadratic fields inside the HilbertClass field. And actually genus theory gives a quite nice description of all quadratic fields in the Hilbert class field. So that the upshot is that for N prime we only need to check if p splits in Q(sqrt( N )) if -N is not a fundamental discriminant, in the case that -N is a fundamental discriminant, there is no extra condition we need to check.